Safely handle missing .mozconfig file in build-gecko.sh

+++ tools/development/build-gecko.sh
@@ -16,7 +16,9 @@ if [ ! -d "$FIREFOX_DIR" ]; then
 	exit 1
 fi
 
-mv "$FIREFOX_DIR/.mozconfig" "$FIREFOX_DIR/.mozconfig.bak"
+if [ -f "$FIREFOX_DIR/.mozconfig" ]; then
+	mv "$FIREFOX_DIR/.mozconfig" "$FIREFOX_DIR/.mozconfig.bak"
+fi
 
 {
 	echo "ac_add_options --enable-application=mobile/ios"
@@ -38,5 +40,7 @@ fi
 cd "$FIREFOX_DIR"
 ./mach build
 
-rm "$FIREFOX_DIR/.mozconfig"
-mv "$FIREFOX_DIR/.mozconfig.bak" "$FIREFOX_DIR/.mozconfig"
+rm -f "$FIREFOX_DIR/.mozconfig"
+if [ -f "$FIREFOX_DIR/.mozconfig.bak" ]; then
+	mv "$FIREFOX_DIR/.mozconfig.bak" "$FIREFOX_DIR/.mozconfig"
+fi
